Updated program to accept column specifications from either pairs or pairs2. Made selection a json entry.
Added a boolean json entry to allow user to opt for either two row or one row entry when sequence start = a and increment = b (insert (a,b)/(b,a) or only (a,b).

Added a python program that calculates sigma2 values for verifying data.

Changes to main.rs
changed the commit to be on the started transaction. IT was on the pool which defeated the desire to only commit every 500 updates

Changed sigma2() function to return a u128 instead of a string in support of changing sigma2 column in oddonlyresults from varchar(75) to numeric.

Changed the table create logic for oddonlyresult to add the sigma2mod10 and sigma2mod100 columns.

cargo.lock was changed by cargo upon the updates noted above.